+++ src/packages/excalidraw/example/App.tsx
@@ -799,7 +799,7 @@ export default function App({ appTitle, useCustom, customArgs }: AppProps) {
                 files: excalidrawAPI.getFiles(),
               });
               const ctx = canvas.getContext("2d")!;
-              ctx.font = "30px Virgil";
+              ctx.font = "30px HandWritingFont";
               ctx.strokeText("My custom text", 50, 60);
               setCanvasUrl(canvas.toDataURL());
             }}