+++ src/scene/export.ts
@@ -141,8 +141,8 @@ export const exportToSvg = async (
   <defs>
     <style class="style-fonts">
       @font-face {
-        font-family: "Virgil";
-        src: url("${assetPath}Virgil.woff2");
+        font-family: "HandWritingFont";
+        src: url("${assetPath}HandWritingFont.woff2");
       }
       @font-face {
         font-family: "Cascadia";