+++ src/tests/data/restore.test.ts
@@ -57,7 +57,7 @@ describe("restoreElements", () => {
     const textElement = API.createElement({
       type: "text",
       fontSize: 14,
-      fontFamily: FONT_FAMILY.Virgil,
+      fontFamily: FONT_FAMILY.HandWritingFont,
       text: "text",
       textAlign: "center",
       verticalAlign: "middle",